requirements.txt
Flask==2.0.1
firebase-admin==5.0.0
requests==2.25.1
bcrypt==3.2.0
pyjwt==2.1.0
flask-cors==3.0.10
gunicorn==20.1.0
